+++ src/app/_components/Audio.tsx
@@ -1,23 +1,9 @@
 'use client';
-import { useContext, useEffect, useRef } from 'react';
+import { useContext, useEffect } from 'react';
 import { Context } from '../Provider';
-import { TRACK_LIST } from '@/constants/constants';
 
 const Audio = () => {
-  const { audioRef, musicSrc, changeMusicSrc, playMusic } = useContext(Context);
-  const trackIndexRef = useRef(0);
-
-  const handleEnded = () => {
-    if (audioRef.current) {
-      trackIndexRef.current += 1;
-      if (trackIndexRef.current < TRACK_LIST.length) {
-        changeMusicSrc(TRACK_LIST[trackIndexRef.current].src);
-      } else {
-        trackIndexRef.current = 0;
-        changeMusicSrc(TRACK_LIST[trackIndexRef.current].src);
-      }
-    }
-  };
+  const { audioRef, musicSrc, playNextTrack, playMusic } = useContext(Context);
 
   useEffect(() => {
     playMusic();
@@ -27,7 +13,7 @@ const Audio = () => {
     <audio
       ref={audioRef}
       src={musicSrc}
-      onEnded={handleEnded}
+      onEnded={playNextTrack}
       className="hidden"
     />
   );
